requirements.txt
django
djangorestframework
drf_spectacular
